Add REST client.
