+++ app/src/test/java/io/apicurio/registry/RegistryClientTest.java
@@ -1,0 +1,37 @@
+/*
+ * Copyright 2019 Red Hat
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.apicurio.registry;
+
+import io.apicurio.registry.client.RegistryClient;
+import io.apicurio.registry.client.RegistryService;
+import io.quarkus.test.junit.QuarkusTest;
+import org.junit.jupiter.api.Test;
+
+/**
+ * @author Ales Justin
+ */
+@QuarkusTest
+public class RegistryClientTest {
+
+    @Test
+    public void testSmoke() throws Exception {
+        try (RegistryService service = RegistryClient.create("http://localhost:8081")) {
+            service.deleteAllGlobalRules();
+        }
+    }
+
+}
