+++ client/src/main/java/io/apicurio/registry/client/RegistryFilter.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright 2019 Red Hat
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package io.apicurio.registry.client;
+
+import java.util.logging.Logger;
+import javax.ws.rs.client.ClientRequestContext;
+import javax.ws.rs.client.ClientRequestFilter;
+import javax.ws.rs.client.ClientResponseContext;
+import javax.ws.rs.client.ClientResponseFilter;
+
+/**
+ * @author Ales Justin
+ */
+public class RegistryFilter implements ClientRequestFilter, ClientResponseFilter {
+
+    private static Logger log = Logger.getLogger(RegistryFilter.class.getName());
+
+    // TODO -- cache
+
+
+    @Override // request filter
+    public void filter(ClientRequestContext context) {
+        log.warning(context.getUri().toString());
+    }
+
+    @Override
+    public void filter(ClientRequestContext requestContext, ClientResponseContext responseContext) {
+    }
+}
