+++ common/src/main/java/io/apicurio/registry/rest/beans/Error.java
