+++ common/src/main/java/io/apicurio/registry/rest/beans/Rule.java
