+++ common/src/main/java/io/apicurio/registry/rest/beans/VersionMetaData.java
@@ -16,12 +16,13 @@
 
 package io.apicurio.registry.rest.beans;
 
-import java.util.Date;
 import com.fasterxml.jackson.annotation.JsonInclude;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.fasterxml.jackson.annotation.JsonPropertyDescription;
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
 import io.apicurio.registry.types.ArtifactType;
+
+import java.util.Date;
 
 
 /**
