+++ common/src/main/java/io/apicurio/registry/types/ArtifactType.java
@@ -16,10 +16,11 @@
 
 package io.apicurio.registry.types;
 
-import java.util.HashMap;
-import java.util.Map;
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonValue;
+
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * WARNING: Ensure backward compatibility when updating the string values,
