+++ common/src/main/resources/META-INF/beans.xml
@@ -1,0 +1,5 @@
+<beans xmlns="http://java.sun.com/xml/ns/javaee"
+       xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+       xsi:schemaLocation="http://java.sun.com/xml/ns/javaee
+  http://java.sun.com/xml/ns/javaee/beans_2_0.xsd">
+</beans>
