+++ pom.xml
@@ -73,7 +73,9 @@
 	</developers>
 
 	<modules>
+		<module>common</module>
 		<module>app</module>
+		<module>client</module>
 		<module>storage/jpa</module>
 		<module>distro</module>
 		<module>tests</module>
@@ -132,7 +134,17 @@
 			<!-- This project -->
 			<dependency>
 				<groupId>io.apicurio</groupId>
+				<artifactId>apicurio-registry-common</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>io.apicurio</groupId>
 				<artifactId>apicurio-registry-app</artifactId>
+				<version>${project.version}</version>
+			</dependency>
+			<dependency>
+				<groupId>io.apicurio</groupId>
+				<artifactId>apicurio-registry-client</artifactId>
 				<version>${project.version}</version>
 			</dependency>
 			<dependency>
